Major refactor: Changed splitTenderCharge.create signature & fixed metadata handling
Split tender charge creation: instead of taking in params to specify the amount to charge both Lightrail and Stripe, only take the amount to charge Lightrail: asking for both is redundant and is now inconsistent with other client libraries. Changed tests to reflect this.

Metadata handling: previously, any metadata passed in was overwritten, and split tender metadata was not being properly applied to the Stripe charge or the Lightrail capture/void transaction.

Also extracted construction of params specific to Lightrail/Stripe to helper methods.